feat(settings): rebuild Settings as unified configuration hub with sidebar nav
Extract the monolithic settings page into dedicated sub-pages (Tools, Billing,
Preferences, API Keys, Account) with a vertical sidebar layout. The root
/settings route redirects to /settings/tools. Existing tools and billing pages
are preserved; profile, notifications, API keys, and danger zone are split
into their own routes with dedicated server loads.
